+++ src/main/java/com/bokmcdok/butterflies/event/entity/living/MobSpawnEventListener.java
@@ -1,5 +1,6 @@
 package com.bokmcdok.butterflies.event.entity.living;
 
+import com.bokmcdok.butterflies.config.ButterfliesConfig;
 import com.bokmcdok.butterflies.registries.EntityTypeRegistry;
 import net.minecraft.world.entity.EntityType;
 import net.minecraft.world.entity.MobSpawnType;
@@ -9,7 +10,10 @@ import net.neoforged.bus.api.IEventBus;
 import net.neoforged.bus.api.SubscribeEvent;
 import net.neoforged.neoforge.event.EventHooks;
 import net.neoforged.neoforge.event.entity.living.MobSpawnEvent;
-
+import net.minecraft.world.entity.monster.AbstractIllager;
+import net.minecraft.world.entity.monster.Witch;
+import net.minecraft.world.entity.npc.Villager;
+import net.minecraft.world.entity.raid.Raider;
 /**
  * Holds event listeners for entities.
  */
@@ -30,14 +34,21 @@ public class MobSpawnEventListener {
     }
 
     /**
+     * Handle mobs being replaced on spawn.
+     * @param event The event context.
+     */
+    private void onMobSpawn(MobSpawnEvent.FinalizeSpawn event) {
+        trySpawnButterflyGolem(event);
+        trySpawnPeacemakerButterfly(event);
+    }
+
+    /**
      * Occasionally replace an iron golem with a butterfly golem.
      * @param event The event context.
      */
     @SubscribeEvent
     @SuppressWarnings({"deprecation", "UnstableApiUsage", "OverrideOnly"})
-    private void onMobSpawn(MobSpawnEvent.FinalizeSpawn event) {
-
-        // Only affect Iron Golems.
+    private void trySpawnButterflyGolem(MobSpawnEvent.FinalizeSpawn event) {
         if (event.getEntity().getType() == EntityType.IRON_GOLEM) {
             IronGolem ironGolem = (IronGolem) event.getEntity();
 
@@ -62,6 +73,52 @@ public class MobSpawnEventListener {
                         }
                     }
                 }
+            }
+        }
+    }
+
+    /**
+     * Some Illagers and Villagers may be infected with Peacemaker Butterflies.
+     * @param event The event context.
+     */
+    private void trySpawnPeacemakerButterfly(MobSpawnEvent.FinalizeSpawn event) {
+
+        // Peacemaker butterflies can be disabled via a config.
+        if (ButterfliesConfig.Common.enableHostileButterflies.get()) {
+            Entity entity = event.getEntity();
+
+            // Handle Villagers being infected.
+            if (entity instanceof Villager villager) {
+                ServerLevelAccessor level = event.getLevel();
+                if (villager.getRandom().nextInt(1000) < 17) {
+                    PeacemakerButterfly.possess(level, villager);
+                    event.setCanceled(true);
+                }
+            }
+
+            // Handle raiders being infected.
+            if (entity instanceof Raider raider) {
+                ServerLevelAccessor level = event.getLevel();
+                if (raider.getRandom().nextInt(100) < 5) {
+                    PeacemakerButterfly.possess(level, raider);
+                    event.setCanceled(true);
+                }
+            }
+        }
+    }
+
+    /**
+     * If a villager, illager, or witch is killed by a peacemaker butterfly
+     * then it shouldn't drop loot.
+     * @param event The drop event to cancel
+     */
+    private void onLivingDrops(LivingDropsEvent event) {
+        if (event.getSource().getEntity() instanceof PeacemakerButterfly) {
+            LivingEntity killed = event.getEntity();
+            if (killed instanceof Villager ||
+                    killed instanceof AbstractIllager ||
+                    killed instanceof Witch) {
+                event.setCanceled(true);
             }
         }
     }